when date is incompatible with year
there are observations in which the date, recorded by CL, doesn't match
the study year, recorded by RAND. Specifically, if the student worked
the problem some time after August 2008, but the student was part of
study year 1. Or if the student was part of study year 2 and worked
problems before July 2008. We removed those problems (34,602)  which
also removed usage data for 178 students.

Also:  retain original curriculum data, and order status variable